feat: revert restructure
for ArraySummary, ButtonNavigation, formFlowContext, width.ts
