+++ app/components/arraySummary/__test__/ArraySummaryDataItems.test.tsx
@@ -23,12 +23,9 @@ const translations = {
   "unterhaltszahlungen.surname": "Nachname",
 };
 
-vi.mock(
-  "~/components/formElements/arraySummary/ArraySummaryItemButton",
-  () => ({
-    default: () => <div>Mock ArraySummaryItemButton</div>,
-  }),
-);
+vi.mock("~/components/arraySummary/ArraySummaryItemButton", () => ({
+  default: () => <div>Mock ArraySummaryItemButton</div>,
+}));
 
 describe("ArraySummaryDataItems", () => {
   afterEach(() => {
