+++ app/components/summaryOverview/__test__/SummaryOverviewBoxArray.test.tsx
@@ -1,10 +1,10 @@
 import { render } from "@testing-library/react";
-import { useFormFlow } from "../../formElements/formFlowContext";
+import { useFormFlow } from "../../formFlowContext";
 import { addArrayIndexToPathUrl } from "../addArrayIndexToPathUrl";
 import { getArraySummaryObject } from "../getArraySummaryObject";
 import SummaryOverviewBoxArray from "../SummaryOverviewBoxArray";
 
-vi.mock("../../formElements/formFlowContext", () => ({
+vi.mock("../../formFlowContext", () => ({
   useFormFlow: vi.fn(),
 }));
