+++ app/routes/beratungshilfe.zustaendiges-gericht.suche.tsx
@@ -7,10 +7,10 @@ import type { ActionFunctionArgs, LoaderFunctionArgs } from "react-router";
 import { data, redirect, useLoaderData } from "react-router";
 import { z } from "zod";
 import Background from "~/components/Background";
+import { ButtonNavigation } from "~/components/ButtonNavigation";
 import Container from "~/components/Container";
 import ContentComponents from "~/components/ContentComponents";
 import { FormComponents } from "~/components/FormComponents";
-import { ButtonNavigation } from "~/components/formElements/ButtonNavigation";
 import { ReportProblem } from "~/components/reportProblem/ReportProblem";
 import type { FlowId } from "~/domains/flowIds";
 import { fetchFlowPage, fetchTranslations } from "~/services/cms/index.server";
