+++ app/util/buttonProps.ts
@@ -1,4 +1,4 @@
-import type { ButtonNavigationProps } from "~/components/formElements/ButtonNavigation";
+import type { ButtonNavigationProps } from "~/components/ButtonNavigation";
 
 export function getButtonNavigationProps({
   backButtonLabel,
